pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snake-drl-agent"
version = "0.1.0"
description = "A Snake game UI with a Deep Reinforcement Learning agent to play the Snake game"
readme = "README.md"
# authors = [
#     {name = "Pradeep",email = ""},
# ]
license = {text = "Apache License 2.0"}
requires-python = ">=3.8"
dependencies = [
    "gym>=0.21.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "torch>=1.12.0",
    "torchvision>=0.13.0",
    "loguru>=0.6.0",
    "streamlit>=1.10.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.21.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.12.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pylint>=2.15.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pylint.messages_control]
max-line-length = 88

[tool.pylint.design]
max-args = 10
max-locals = 15

[tool.pylint.format]
max-line-length = 88